Move QueryPlanSerializationSettings to Settings.cpp
