+++ src/Processors/QueryPlan/ArrayJoinStep.cpp
@@ -1,5 +1,5 @@
 #include <Processors/QueryPlan/ArrayJoinStep.h>
-#include <Processors/QueryPlan/QueryPlanSerializationSettings.h>
+#include <Core/QueryPlanSerializationSettings.h>
 #include <Processors/QueryPlan/QueryPlanStepRegistry.h>
 #include <Processors/QueryPlan/Serialization.h>
 #include <Processors/Transforms/ArrayJoinTransform.h>
